More Commenting
Commented Up to PageController, also new idea for deleting entries which have other related entries, set ID to 0 and have defaulted/unset ID for those tables.

+++ app/Http/Controllers/JobController.php
@@ -19,10 +19,13 @@ class JobController extends Controller
      */
     public function index()
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {
+            // Get all jobs, with their department names, and number of employees.
             $jobs = Department::join('jobs', 'departments.id', '=', 'jobs.department_id')->leftJoin('users', 'jobs.id', '=', 'users.job_id')->selectRaw('jobs.id, jobs.title, departments.name, departments.id as dID, IFNULL(COUNT(users.id),0) as employees')->groupBy('jobs.id')->get();
 
+            // Supply data to view.
             $data = array(
                 'title' => "Job Information Viewer",
                 'desc' => "View information on jobs.",
@@ -30,10 +33,10 @@ class JobController extends Controller
                 'links'=>PagesController::getOperatorLinks(),
                 'active'=>'Jobs'
             );
-
             return view('jobs.index')->with($data);
         }
 
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -44,23 +47,25 @@ class JobController extends Controller
      */
     public function create()
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {
+            // Get all departments.
             $departments = Department::all();
-
-            $dept = $departments->first();
-
-            // If specified, grab the specified department,
-            // for the dropdown box.
-            if (isset($_GET['department']) && !empty($_GET['department']))
+            $dept = null;
+            // If supplied, pass given department to view for dropdown box.
+            if (isset($_GET['department']))
             {
                 $dept = Department::find($_GET['department']);
-                if (is_null($dept))
-                {
-                    $dept = $departments->first();
-                }
             }
 
+            // By default use the first in the table.
+            if (is_null($dept))
+            {
+                $dept = $departments->first();
+            }
+
+            // Supply data to view.
             $data = array(
                 'title' => "Create New Job",
                 'desc' => "For making a new job title.",
@@ -72,6 +77,7 @@ class JobController extends Controller
 
             return view('jobs.create')->with($data);
         }
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -83,26 +89,33 @@ class JobController extends Controller
      */
     public function store(Request $request)
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {            
+            // Validate that the required fields have been supplied.
             $this->validate($request, [
                 'department-select' => 'required',
                 'jobTitle' => 'required'
             ]);
 
+            // Find jobs with matching title as new title.
             $job = Job::where('title', $request->input('jobTitle'))->get();
-
-            if (count($job) == 0)
+            $department = Department::find($request->input('department-select'));
+            // No matching jobs.
+            if (count($job) == 0 && !is_null($department))
             {
+                // Create job entry, redirect to department job is found in.
                 $newJob = new Job();
                 $newJob->title = $request->input('jobTitle');
-                $newJob->department_id = $request->input('department-select');
+                $newJob->department_id = $department->id;
+                // Only the technical support jobs can have access level != 0, and they are immutable.
                 $newJob->access_level = '0';
                 $newJob->save();
 
-                return redirect('/departments/'.$request->input('department-select'))->with('success', 'Job Added');
+                return redirect('/departments/'.$department->id)->with('success', 'Job Added');
             }
 
+            // If matching job(s) are found.
             $data = array(
                 'error'=>'Duplicate Job Name',
                 'search'=>$request->input('jobTitle')
@@ -110,6 +123,7 @@ class JobController extends Controller
 
             return redirect('/jobs')->with($data);
         }
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -121,13 +135,22 @@ class JobController extends Controller
      */
     public function show($id)
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {
+            // Find relevant job from ID.
             $job = Job::find($id);
-            $department = Department::where('id', $job->department_id)->get()->first();
+            if (is_null($job))
+            {
+                return redirect()->back();
+            }
 
+            // Find out which department job is part of.
+            $department = Department::find($job->department_id);
+            // Get all users who have this job.
             $users = DB::table('users')->join('jobs', 'users.job_id', '=', 'jobs.id')->join('departments', 'jobs.department_id', '=', 'departments.id')->select('users.*', 'jobs.access_level', 'departments.name')->where('jobs.id', '=', $id)->get();
 
+            // Supply data to view.
             $data = array(
                 'title' => $job->title,
                 'desc' => "Display information on a job.",
@@ -141,6 +164,7 @@ class JobController extends Controller
             return view('jobs.show')->with($data);
         }
 
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -152,27 +176,32 @@ class JobController extends Controller
      */
     public function edit($id)
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {
+            // Find relevant job from ID.
             $job = Job::find($id);
-            $dept = Department::find($job->department_id);
-            if (!is_null($job))
+            if (is_null($job))
             {
-                $departments = Department::where('id', '!=', '1')->get();
-                $data = array(
-                    'title' => "Edit Existing Job",
-                    'desc' => "For making a new department catagory.",
-                    'job'=>$job,
-                    'dept'=>$dept,
-                    'departments'=>$departments,
-                    'links' => PagesController::getOperatorLinks(),
-                    'active' => 'Jobs'
-                );
-
-                return view('jobs.edit')->with($data);
+                return redirect()->back();
             }
-            return redirect('/jobs');
+            $dept = Department::find($job->department_id);
+            // Get all department info except tech support.
+            $departments = Department::where('id', '!=', '1')->get();
+            // Supply data to view.
+            $data = array(
+                'title' => "Edit Existing Job",
+                'desc' => "For making a new department catagory.",
+                'job'=>$job,
+                'dept'=>$dept,
+                'departments'=>$departments,
+                'links' => PagesController::getOperatorLinks(),
+                'active' => 'Jobs'
+            );
+
+            return view('jobs.edit')->with($data);
         }
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -185,25 +214,37 @@ class JobController extends Controller
      */
     public function update(Request $request, $id)
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {            
+            // Validate that the required fields have been supplied.
             $this->validate($request, [
                 'jobTitle' => 'required',
                 'department-select' => 'required'
             ]);
 
+            // Find matching jobs that already have same title as one supplied.
             $job = Job::where('title', '=', $request->input('jobTitle'))->where('id', '!=', $id)->get();
+            // No matches found.
             if (count($job) == 0)
             {
-                $newJob = Job::find($id);
-                $newJob->title = $request->input('jobTitle');
-                $newJob->department_id = $request->input('department-select');
-                $newJob->access_level = '0';
-                $newJob->save();
+                // Find job that needs to be edited.
+                $job = Job::find($id);
+                $department = Department::find($request->input('department-select'));
+                // If job/department ID doesn't exist, redirect.
+                if (is_null($job) || is_null($department))
+                {
+                    return redirect('/jobs')->with('error', 'Invalid data supplied.');
+                }
+                // Update info for title and save to database.
+                $job->title = $request->input('jobTitle');
+                $job->department_id = $department;
+                $job->save();
 
                 return redirect('/jobs')->with('success', 'Job Updated');
             }
 
+            // Supply data to view.
             $data = array(
                 'error'=>'Duplicate Job Name',
                 'search'=>$request->input('jobTitle')
@@ -211,6 +252,7 @@ class JobController extends Controller
 
             return redirect('/jobs')->with($data);
         }
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 
@@ -222,15 +264,25 @@ class JobController extends Controller
      */
     public function destroy($id)
     {
+        // If user can access page
         if (PagesController::hasAccess(1))
         {
             $job = Job::find($id);
+            if (is_null($job))
+            {
+                return redirect()->back();
+            }
             $job->delete();
 
-            $users = User::where('job_id', $id)->delete();
+            $users = User::where('job_id', $id)->get();
+            foreach ($users as $u) {
+                $u->job_id = 0;
+                $u->save();
+            }
 
             return redirect('/jobs')->with('success', 'Job Deleted');
         }
+        // No access redirects to login.
         return redirect('login')->with('error', 'Please log in first.');
     }
 }
